backend/requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
Flask-SQLAlchemy>=3.1.0
PyYAML>=6.0
pandas>=2.0.0
yfinance>=0.2.28
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

